requirements.txt
appium-python-client
pytest
allure-pytest
pytesseract
Pillow
openai  # or your chosen LLM wrapper
